[refactor] 拆分 user controllers 結構
